requirements.txt
streamlit
numpy
tensorflow
opencv-python
SimpleITK
Pillow
matplotlib
fpdf